Isolated Speech Recognition using LCNN
The aim of the project is to classify and evaluate spoken digits (0-9) by the help of Signal Processing for feature extraction and Machine Learning for Classification. By using Convolutional Neural Network (LCNN), Rectifier Linear unit (ReLU) activation function is not used but rather Max-Feature-Map (MFM) is used for faster classification and better results. Managed to have the accuracy between 70% and 95%.